+++ src/api/entity.hpp
@@ -141,13 +141,13 @@ namespace copper_server::api {
 
             void (*entity_attack)(ecs::entity self, ecs::entity, ecs::entity other_entity_id) = nullptr;
             void (*entity_iteract)(ecs::entity self, ecs::entity, ecs::entity other_entity_id) = nullptr;
-            void (*entity_iteract_block)(ecs::entity self, ecs::entity, int64_t x, int64_t y, int64_t z) = nullptr;
+            void (*entity_iteract_block)(ecs::entity self, ecs::entity, int32_t x, int32_t y, int32_t z) = nullptr;
 
-            void (*entity_break)(ecs::entity self, ecs::entity, int64_t x, int64_t y, int64_t z, uint8_t state) = nullptr; //form 0 to 9, other ignored
-            void (*entity_cancel_break)(ecs::entity self, ecs::entity, int64_t x, int64_t y, int64_t z) = nullptr;
-            void (*entity_finish_break)(ecs::entity self, ecs::entity, int64_t x, int64_t y, int64_t z) = nullptr;
-            void (*entity_place_block)(ecs::entity self, ecs::entity, bool is_main_hand, int64_t x, int64_t y, int64_t z, const base_objects::block&) = nullptr;
-            void (*entity_place_block_entity)(ecs::entity self, ecs::entity, bool is_main_hand, int64_t x, int64_t y, int64_t z, ecs::entity) = nullptr;
+            void (*entity_break)(ecs::entity self, ecs::entity, int32_t x, int32_t y, int32_t z, uint8_t state) = nullptr; //form 0 to 9, other ignored
+            void (*entity_cancel_break)(ecs::entity self, ecs::entity, int32_t x, int32_t y, int32_t z) = nullptr;
+            void (*entity_finish_break)(ecs::entity self, ecs::entity, int32_t x, int32_t y, int32_t z) = nullptr;
+            void (*entity_place_block)(ecs::entity self, ecs::entity, bool is_main_hand, int32_t x, int32_t y, int32_t z, const base_objects::block&) = nullptr;
+            void (*entity_place_block_entity)(ecs::entity self, ecs::entity, bool is_main_hand, int32_t x, int32_t y, int32_t z, ecs::entity) = nullptr;
 
 
             void (*entity_animation)(ecs::entity self, ecs::entity, base_objects::entity_animation animation) = nullptr;
@@ -160,21 +160,21 @@ namespace copper_server::api {
             void (*entity_deinit)(ecs::entity self, ecs::entity) = nullptr;
 
 
-            void (*notify_block_event)(ecs::entity self, const base_objects::world::block_action& action, int64_t x, int64_t y, int64_t z) = nullptr;
-            void (*notify_block_change)(ecs::entity self, int64_t x, int64_t y, int64_t z, const base_objects::block& block) = nullptr;
-            void (*notify_block_entity_change)(ecs::entity self, int64_t x, int64_t y, int64_t z, ecs::entity) = nullptr;
-            void (*notify_block_destroy_change)(ecs::entity self, int64_t x, int64_t y, int64_t z, const base_objects::block& block) = nullptr;
-            void (*notify_block_entity_destroy_change)(ecs::entity self, int64_t x, int64_t y, int64_t z, ecs::entity) = nullptr;
-            void (*notify_biome_change)(ecs::entity self, int64_t x, int64_t y, int64_t z, uint32_t) = nullptr;
+            void (*notify_block_event)(ecs::entity self, const base_objects::world::block_action& action, int32_t x, int32_t y, int32_t z) = nullptr;
+            void (*notify_block_change)(ecs::entity self, int32_t x, int32_t y, int32_t z, const base_objects::block& block) = nullptr;
+            void (*notify_block_entity_change)(ecs::entity self, int32_t x, int32_t y, int32_t z, ecs::entity) = nullptr;
+            void (*notify_block_destroy_change)(ecs::entity self, int32_t x, int32_t y, int32_t z, const base_objects::block& block) = nullptr;
+            void (*notify_block_entity_destroy_change)(ecs::entity self, int32_t x, int32_t y, int32_t z, ecs::entity) = nullptr;
+            void (*notify_biome_change)(ecs::entity self, int32_t x, int32_t y, int32_t z, uint32_t) = nullptr;
 
-            void (*notify_sub_chunk)(ecs::entity self, int64_t chunk_x, int64_t chunk_y, int64_t chunk_z, const base_objects::world::sub_chunk_data&) = nullptr; //used after multiply changes
-            void (*notify_chunk)(ecs::entity self, int64_t chunk_x, int64_t chunk_z, const base_objects::world::chunk_data&) = nullptr;                          //used after multiply changes
+            void (*notify_sub_chunk)(ecs::entity self, int32_t chunk_x, int32_t chunk_y, int32_t chunk_z, const base_objects::world::sub_chunk_data&) = nullptr; //used after multiply changes
+            void (*notify_chunk)(ecs::entity self, int32_t chunk_x, int32_t chunk_z, const base_objects::world::chunk_data&) = nullptr;                          //used after multiply changes
 
-            void (*notify_sub_chunk_light)(ecs::entity self, int64_t chunk_x, int64_t chunk_y, int64_t chunk_z, const base_objects::world::sub_chunk_data&) = nullptr; //used after multiply changes
-            void (*notify_chunk_light)(ecs::entity self, int64_t chunk_x, int64_t chunk_z, const base_objects::world::chunk_data&) = nullptr;                          //used after multiply changes
+            void (*notify_sub_chunk_light)(ecs::entity self, int32_t chunk_x, int32_t chunk_y, int32_t chunk_z, const base_objects::world::sub_chunk_data&) = nullptr; //used after multiply changes
+            void (*notify_chunk_light)(ecs::entity self, int32_t chunk_x, int32_t chunk_z, const base_objects::world::chunk_data&) = nullptr;                          //used after multiply changes
 
-            void (*notify_sub_chunk_blocks)(ecs::entity self, int64_t chunk_x, int64_t chunk_y, int64_t chunk_z, const base_objects::world::sub_chunk_data&) = nullptr; //used after multiply changes
-            void (*notify_chunk_blocks)(ecs::entity self, int64_t chunk_x, int64_t chunk_z, const base_objects::world::chunk_data&) = nullptr;                          //used after multiply changes
+            void (*notify_sub_chunk_blocks)(ecs::entity self, int32_t chunk_x, int32_t chunk_y, int32_t chunk_z, const base_objects::world::sub_chunk_data&) = nullptr; //used after multiply changes
+            void (*notify_chunk_blocks)(ecs::entity self, int32_t chunk_x, int32_t chunk_z, const base_objects::world::chunk_data&) = nullptr;                          //used after multiply changes
 
             void (*on_change_world)(ecs::entity self, storage::world_data& new_world) = nullptr;
 
@@ -323,9 +323,9 @@ namespace copper_server::api {
         //attack passes only if this and another entity registered to same world
         void attack_from_this(ecs::entity entity);
 
-        void breaking_block(int64_t global_x, uint64_t global_y, int64_t global_z, uint32_t time);
-        void place_block(int64_t global_x, uint64_t global_y, int64_t global_z, const base_objects::block&);
-        void place_block(int64_t global_x, uint64_t global_y, int64_t global_z, ecs::entity block_entity);
+        void breaking_block(int32_t global_x, int32_t global_y, int32_t global_z, uint32_t time);
+        void place_block(int32_t global_x, int32_t global_y, int32_t global_z, const base_objects::block&);
+        void place_block(int32_t global_x, int32_t global_y, int32_t global_z, ecs::entity block_entity);
 
         static ecs::entity create(int32_t id);
         static ecs::entity create(int32_t id, const util::nbt_compound& nbt);
